requirements.txt
streamlit
akshare
openai
python-dotenv
pandas
plotly
faiss-cpu
numpy
